fix duplication of solution versions and add comparison solutions

+++ app/Http/Controllers/User/SolutionController.php
@@ -17,12 +17,16 @@ class SolutionController extends Controller
 
     public function __construct(ActivityService $activityService)
     {
-        $this->authorizeResource(Solution::class, 'solution');
+//        $this->authorizeResource(Solution::class, 'solution', [
+//            'except' => ['show'],
+//        ]);
         $this->activityService = $activityService;
     }
 
     public function store(Request $request, User $user): RedirectResponse
     {
+        $this->authorize('create', Solution::class);
+
         $validatedData = $request->validate([
             'content' => 'required|string|min:1',
         ]);
@@ -45,23 +49,30 @@ class SolutionController extends Controller
 
     public function show(User $user, Solution $solution): View
     {
-
         $currentExercise = $solution->exercise;
 
-        $solutionsListForCurrentExercise = $solution->exercise
+        $solutionsOfCurrentUser = $solution->exercise
         ->solutions()
         ->where('user_id', $user->id)
         ->get();
 
-        return view('solution.show', compact(
+        $solutionsOfOtherUsers = $solution->exercise
+            ->solutions()
+            ->whereNot('user_id', $user->id)
+            ->get();
+
+        return view('user.solution.show', compact(
             'currentExercise',
-            'solutionsListForCurrentExercise',
-            'user'
+            'solutionsOfCurrentUser',
+            'solutionsOfOtherUsers',
+            'user',
         ));
     }
 
     public function destroy(User $user, Solution $solution): RedirectResponse
     {
+        $this->authorize('delete', $solution);
+
         if ($solution->delete()) {
             flash()->success(__('layout.flash.success'));
         } else {
